뽀모도로 관련 공유 타입 및 유틸을 최상위로 이동
