+++ src/renderer/features/pomodoro/index.ts
@@ -1,2 +1,3 @@
 export * from './hooks/use-add-pomodoro';
+export * from './hooks/use-pomodoro-by-main';
 export * from './util/cat-name-map';
