+++ src/renderer/shared/utils/index.ts
@@ -2,7 +2,5 @@ export * from './cn';
 export * from './url';
 export * from './storage';
 export * from './icon';
-export * from './string';
-export * from './time';
 export * from './iso-duration';
 export * from './datadog';
