--- src/renderer/shared/utils/string.ts
@@ -1,1 +1,0 @@
-export const padNumber = (num: number) => num.toString().padStart(2, '0');
