+++ src/renderer/tsconfig.json
@@ -17,6 +17,7 @@
     "outDir": "dist",
     "baseUrl": ".",
     "paths": {
+      "shared/*": ["../shared/*"],
       "@/*": ["./*"]
     }
   }
