+++ src/shared/type/index.ts
@@ -1,0 +1,2 @@
+export * from './electron';
+export * from './pomodoro';
