+++ src/shared/util/string.ts
